+++ plugins/search_engine_lookup/docs/HISTORY.md
@@ -12,6 +12,15 @@ The following people have contributed to the development of this plugin.
 
 The following identifies the development history of the plugin, in reverse chronological order.  Each version lists the changes made for that version, along with the author of each change.
 
+### Version 1.0.0
+
+* Initial release submitted to the Picard Plugins listing.
+
+### Version 0.0.3
+
+* Add option setting to provide additional words to use with the search. \[rdswift\]
+* Add pop-up dialog to notify if an error occurred preventing the search. \[rdswift\]
+
 ### Version 0.0.2
 
 * Add option settings to select the search engine provider to use. \[rdswift\]